+++ src/amun/gpu/dl4mt/cellstate.h
@@ -18,7 +18,7 @@ struct CellState {
   std::unique_ptr<mblas::Matrix> output;
   std::unique_ptr<mblas::Matrix> cell;
 
-  std::string Debug(size_t verbosity = 1) const;
+  std::string Debug(unsigned verbosity = 1) const;
 
 };
 }
